Refactoring: moved a chess independent notion to the proper package.
