+++ src/main/java/com/leokom/games/chess/player/legal/brain/normalized/ValidatingNormalizedEvaluator.java
@@ -1,7 +1,7 @@
 package com.leokom.games.chess.player.legal.brain.normalized;
 
 import com.leokom.games.chess.engine.GameState;
-import com.leokom.games.chess.engine.GameTransition;
+import com.leokom.games.commons.engine.GameTransition;
 import com.leokom.games.chess.player.legal.brain.common.GenericEvaluator;
 
 /**
