+++ src/main/java/com/leokom/games/commons/engine/GameTransition.java
@@ -1,9 +1,8 @@
-package com.leokom.games.chess.engine;
+package com.leokom.games.commons.engine;
 
 /**
  * The notion of game transition is an attempt to represent the game
  * as a state automate.
- * This notion is so generic that can be extracted to something chess-independent
  */
 public interface GameTransition {
 }
